# docs/swagger.yaml
basePath: /atenea
definitions:
  models.EquipoParam:
    properties:
      descripcion:
        type: string
      id:
        type: integer
      ip:
        type: string
      nombre:
        type: string
    type: object
  models.GrupoEquipoParam:
    properties:
      id_equipo:
        type: integer
      id_grupo:
        type: integer
    type: object
  models.GrupoParam:
    properties:
      id:
        type: integer
      nombre:
        type: string
    type: object
...
  models.MonitoreoEquipo:
    properties:
      critical:
        type: integer
      equipo:
        type: string
      id_equipo:
        type: integer
      ip:
        type: string
      major:
        type: integer
      not_reachable:
        type: integer
      notify:
        type: integer
      warning:
        type: integer
    type: object
  models.MonitoreoGrupo:
    properties:
      critical:
        type: integer
      grupo:
        type: string
      id_grupo:
        type: integer
      major:
        type: integer
      not_reachable:
        type: integer
      notify:
        type: integer
      tot_equipos:
        type: integer
      warning:
        type: integer
    type: object
...
host: IPServer:3021
info:
  contact:
    email: rgarciaxxx@yahoo.com.mx
    name: 'Soporte para las APIS: Raúl García'
    url: http://www.ineel.mx
  description: Contiene las APIs para atender las peticiones de la IHM hacia la base
    de datos ATENEA
  license:
    name: Atenea Inc.
    url: http://www.atenea.mx/LICENSE-2.0.html
  title: Servidor de APIs para ATENEA
  version: "1.0"
paths:
...
  /equipo:
    post:
      consumes:
      - application/json
      description: Agrega un nueva equipo
      parameters:
      - description: Estructura del equipo a agregar
        in: body
        name: equipo
        required: true
        schema:
          $ref: '#/definitions/models.EquipoParam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: ""
        "409":
          description: Conflict
          schema:
            type: string
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Crea un nuevo registro de equipo
      tags:
      - equipo
    put:
      consumes:
      - application/json
      description: Actualiza el registro del equipo indicado en el body de la petición
      parameters:
      - description: Estructura del equipo a actualizar
        in: body
        name: equipo
        required: true
        schema:
          $ref: '#/definitions/models.EquipoParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: ""
        "409":
          description: Conflict
          schema:
            type: string
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Actualiza datos de un equipo
      tags:
      - equipo
  /equipo/id/{id}:
    delete:
      consumes:
      - application/json
      description: Borra el equipo correspondiente al valor del identificador (id)
        indicado.
      parameters:
      - description: Identificador (id) del equipo a eliminar
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: ""
        "409":
          description: Conflict
          schema:
            type: string
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Elimina equipos
      tags:
      - equipo
  /grupo:
    post:
      consumes:
      - application/json
      description: Agrega un nuevo grupo
      parameters:
      - description: Estructura del grupo a agregar
        in: body
        name: equipo
        required: true
        schema:
          $ref: '#/definitions/models.GrupoParam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: ""
        "409":
          description: Conflict
          schema:
            type: string
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Crea un nuevo registro de grupo
      tags:
      - grupo
    put:
      consumes:
      - application/json
      description: Actualiza el registro del grupo indicado en el body de la petición
      parameters:
      - description: Estructura del grupo a actualizar
        in: body
        name: equipo
        required: true
        schema:
          $ref: '#/definitions/models.GrupoParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: ""
        "409":
          description: Conflict
          schema:
            type: string
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Actualiza datos de un grupo
      tags:
      - grupo
  /grupo-equipo:
    post:
      consumes:
      - application/json
      description: Agrega un nuevo grupo-equipo
      parameters:
      - description: Estructura del grupo-equipo a agregar
        in: body
        name: equipo
        required: true
        schema:
          $ref: '#/definitions/models.GrupoEquipoParam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: ""
        "409":
          description: Conflict
          schema:
            type: string
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Crea un nuevo registro de grupo-equipo
      tags:
      - grupo-equipo
  /grupo-equipo/{id_grupo}/{id_equipo}:
    delete:
      consumes:
      - application/json
      description: Borra la combinación grupo - equipo correspondiente a los valores
        de los identificadores (id-equipo, id_grupo) indicado.
      parameters:
      - description: Identificador del grupo
        in: path
        name: id_grupo
        required: true
        type: string
      - description: Identificador del equipo
        in: path
        name: id_equipo
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: ""
        "409":
          description: Conflict
          schema:
            type: string
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Elimina combinaciones de grupos - equipos
      tags:
      - grupo-equipo
  /grupo/id/{id}:
    delete:
      consumes:
      - application/json
      description: Borra el grupo correspondiente al valor del identificador (id)
        indicado.
      parameters:
      - description: Identificador (id) del grupo a eliminar
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: ""
        "409":
          description: Conflict
          schema:
            type: string
        "417":
          description: Expectation Failed
          schema:
            type: string
        "500":
          description: ""
      summary: Elimina grupos
      tags:
      - grupo
  /monitoreo/equipo:
    get:
      description: Obtiene el monitoreo de todos los equipos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MonitoreoEquipo'
            type: array
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene el monitoreo por equipos
      tags:
      - monitoreo
  /monitoreo/equipo/{id}:
    get:
      description: Obtiene el monitoreo del equipo correspondiente al identificador
        del equipo indicado
      parameters:
      - description: Identificador (id) del equipo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MonitoreoEquipo'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene el monitoreo por equipos
      tags:
      - monitoreo
  /monitoreo/grupo/{id}:
    get:
      description: Obtiene el monitoreo del equipo correspondiente al identificador
        del grupo indicado
      parameters:
      - description: Identificador (id) del grupo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MonitoreoEquipo'
            type: array
        "400":
          description: ""
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene el monitoreo por grupos
      tags:
      - monitoreo
  /monitoreo/grupos:
    get:
      description: Obtiene los datos de monitoreo de los grupos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MonitoreoGrupo'
            type: array
        "404":
          description: ""
        "500":
          description: ""
      summary: Obtiene los datos de monitoreo de los grupos
      tags:
      - monitoreo
...
swagger: "2.0"